feat: when bubble click, go to detail screen
