+++ feature/dashboard/src/main/kotlin/com/chipichipi/dobedobe/feature/dashboard/DashboardScreen.kt
@@ -126,7 +126,6 @@ private fun DashboardScreen(
                     navigateToSetting = navigateToSetting,
                     navigateToSearchGoal = navigateToSearchGoal,
                     onGoalToggled = onGoalToggled,
-                    // TODO: 캐릭터 누르면 말풍선 바뀌는걸로 옮기기!
                     onChangeBubble = onChangeBubble,
                     onToggleMode = onToggleMode,
                     onUpsertPhotos = onUpsertPhotos,
@@ -217,11 +216,12 @@ private fun DashboardBody(
                 DashboardViewMode(
                     isViewMode = !isEditMode,
                     photoState = uiState.photoState,
-                    bubbleTitle = uiState.bubbleTitle,
+                    bubbleGoal = uiState.bubbleGoal,
                     photoFramesState = photoFramesState,
                     onChangeBubble = onChangeBubble,
                     onToggleExpansion = onToggleExpansion,
                     onToggleMode = onToggleMode,
+                    navigateToGoalDetail = navigateToGoalDetail,
                     navigateToSetting = navigateToSetting,
                     character = uiState.character,
                 )
