+++ feature/dashboard/src/main/kotlin/com/chipichipi/dobedobe/feature/dashboard/component/DashboardViewMode.kt
@@ -19,6 +19,7 @@ import androidx.compose.ui.zIndex
 import com.chipichipi.dobedobe.core.designsystem.theme.DobeDobeTheme
 import com.chipichipi.dobedobe.core.model.CharacterType
 import com.chipichipi.dobedobe.feature.dashboard.DashboardPhotoFramesState
+import com.chipichipi.dobedobe.feature.dashboard.model.BubbleGoal
 import com.chipichipi.dobedobe.feature.dashboard.model.CharacterResources
 import com.chipichipi.dobedobe.feature.dashboard.model.DashboardPhotoState
 
@@ -27,12 +28,13 @@ import com.chipichipi.dobedobe.feature.dashboard.model.DashboardPhotoState
 internal fun SharedTransitionScope.DashboardViewMode(
     isViewMode: Boolean,
     photoState: List<DashboardPhotoState>,
-    bubbleTitle: String,
+    bubbleGoal: BubbleGoal,
     photoFramesState: DashboardPhotoFramesState,
     onChangeBubble: () -> Unit,
     onToggleExpansion: (Int) -> Unit,
     onToggleMode: () -> Unit,
     navigateToSetting: () -> Unit,
+    navigateToGoalDetail: (Long) -> Unit,
     character: CharacterType,
     modifier: Modifier = Modifier,
 ) {
@@ -76,13 +78,17 @@ internal fun SharedTransitionScope.DashboardViewMode(
                 ) {
                     Spacer(Modifier.height(14.dp))
                     DashboardBubble(
-                        title = bubbleTitle,
+                        title = bubbleGoal.title,
                         textStyle = DobeDobeTheme.typography.body2,
                         modifier = Modifier
                             .background(
                                 color = DobeDobeTheme.colors.white,
                             ),
-                        onClick = onChangeBubble,
+                        onClick = {
+                            if (bubbleGoal.id != null) {
+                                navigateToGoalDetail(bubbleGoal.id)
+                            }
+                        },
                     )
                     DashboardCharacter(
                         modifier = Modifier
@@ -91,6 +97,7 @@ internal fun SharedTransitionScope.DashboardViewMode(
                         defaultApngRes = resources.defaultApngRes,
                         reactionApngRes = resources.reactionApngRes,
                         placeholder = resources.placeholderRes,
+                        onChangeBubble = onChangeBubble,
                     )
                 }
             }
